Updated enemybomb.cs
- changed vol from int to float

+++ Assets/Scripts/enemyBomb.cs
@@ -17,7 +17,7 @@ public class enemyBomb : MonoBehaviour, IDamage
     [SerializeField] int explosionRange;
     [SerializeField] AudioSource aud;
 
-    [SerializeField] int volume;
+    [SerializeField] float volume;
 
     public AudioClip explodeSound;
     public AudioClip beepSound;
